Fix Bug Schedule Not Updating
`update_schedule` was only checking if now was equal to old_week, this
will never run if the day is missed. It now checks if `old_week` is less
than or equal to now. Even though these are string comparisons, the dates
are structured largest to smallest and zero padded so the comparison
still works. The `next_week` value was also being incorrectly set to the
week that just passed so it wasn't being updated properly.