Add hyperlinks to tags list
Tags are hyperlinked to their respective tag urls to make it so you
don't have to click thru to the comic, then the tag link on the page.

This also required chaning the embed layout so tags'll display above
the image because hyperlinks aren't allowed in the footer. The date
is staying down there. Also, because it's "field" it has to have a
name, so the alt text is moving from the header description to the
field name, making it bolded.