Add edit message ability
File specifies the message id (mid) to edit. If it's a regular text message, `text` field
is what the new message should say. If it is an embed, `title`, `url`, `image`, `footer`
can be included if those values should be changed. A `field` dict with `value` and/or
`name` to change the first field.